+++ src/main/java/jalau/at18/kingoftokyo/view/PlayerCards/PlayerCard.java
@@ -1,4 +1,4 @@
-package jalau.at18.kingoftokyo.view;
+package jalau.at18.kingoftokyo.view.PlayerCards;
 
 import java.awt.Color;
 import javax.swing.JPanel;
