+++ src/main/java/jalau/at18/kingoftokyo/view/PlayerCards/PlayerCardVictoryPoints.java
@@ -1,4 +1,4 @@
-package jalau.at18.kingoftokyo.view;
+package jalau.at18.kingoftokyo.view.PlayerCards;
 
 import java.awt.Font;
 import java.awt.Color;
